Python/requirements.txt
pyrebase4
setuptools
